Combined providers should never return 404
storetheindex never returns 404 on GET to `/providers` even when there
are no providers. Instead, it returns an empty JSON list. Do the same
in indexstar.

Note that this was never happening anyway due to the condition that
was never true; the `resp` was being checked for nil instead of len
zero.