Cap max wait time for results and refine reframe aggregation logic
Implement a reusable scatter-gather task executor which allows the
caller to specify the maximum wait time for gathering results if any.

Refactor the reframe logic to use the scatter-gather executor with
configurable max wait time, default of 5 seconds.

Refine the reframe aggregation logic as following:
* Wait at most `config.Reframe.ResultMaxWait` for a result from backend.
* Only return a single result per peer ID picking the first one found.
* Do not return error if at least one provider is found.
* Return the last observed error if no providers are found.
* If no providers are found and no error has occurred return an empty
  reframe result.

Returning an empty result is consistent with the behaviour of reframe
server implementation in storetheindex.