Implement support for `ndjson` streaming response on IPNI find APIs
Check the `Accept` request header and if it explicitly states support
for ndjson media type, then respond with streaming provider results.

Check uniqueness of results with first-found strategy, where a result
that has already been sent back will not be included again. The check
for uniqueness is done by hashing the provider ID and context ID of a
record using IEEE crc32.

Fixes #65