Implement non-streaming translation to reduce 404s on dht cascading
DHT lookups take longer to complete in comparison with IPNI search. This
is why streaming response was introduced such that results are returned
as soon as they are found instead of having to wait for all results
before returning any.

Internally, indexstar uses a circuit breaker mechanism that would
exclude slow backends from the search space. The max wait for
non-streaming requests is set to 5 seconds. When a client uses
non-streaming `Accept` header with DHT lookup cascade enabled, there is
a high chance that the response would take longer than the internal
timeout of 5 seconds, which results in the DHT cascade endpoints to
eventually be excluded from the search space.

To avoid this, the changes here introduce an option to always translate
non-streaming requests to streaming ndjson, and aggregate the result in
a batch response to the client with internal timeout of 5 seconds. This
would avoid the slow DHT lookups to open the circuit and ultimately
reduce the chances of 404 for the clients that use non-streaming
`Accept` header with dht cascade enabled.